Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bcrypt==3.2.2
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.14
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
markupsafe==3.0.2
motor==3.7.1
openai==1.98.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.20
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
